requirements.txt
django==1.11.4
django-extensions==1.8.1
djangorestframework==3.6.3
doctest2==0.1
gunicorn
requests
jsonfield
#ngrok-client
# pillow
# tensorflow
# keras
# pandas
# scipy
# pugnlp==0.0.9
# git+https://github.com/totalgood/pugnlp.git@master#egg=pugnlp
Pillow==4.2.1
